# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: We're Nerds Dude!! by The Powder Rangers
# during dev, should point to your local machine
host: localhost:10000
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:

  /exercises:
    # binds a127 app logic to a route
    x-swagger-router-controller: exercises
    get:
      description: GET request returns array of all predefined exercises
      # used as the method name of the controller
      operationId: getExercises
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
            # a pointer to a definition
             $ref: "#/definitions/ExercisesResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
  /exercises/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: exercises_id
    get:
      description: GET request returns specific exercise
      # used as the method name of the controller
      operationId: getExercisesId
      parameters:
        - name: id
          in: path
          required: true
          type: integer
          description: Input exercise id
          
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ExercisesResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /routines/:
    x-swagger-router-controller: routines
    get:
      description: GET request returns array of all predefined routines
      operationId: getRoutines
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/RoutinesResponse'
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /routines/{id}:
    x-swagger-router-controller: routines_id
    get:
      description: GET request returns specific predefined routine
      operationId: getRoutinesId
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: Enter routine id
          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RoutinesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /routines/{id}/exercises:
    x-swagger-router-controller: routines_id_exercises
    get:
      description: GET request returns array of exercises for specific routine
      operationId: getRoutinesIdExercises
      parameters:
        - name: id
          in: path
          required: true
          type: string
          description: Enter routine id
          
      responses:
        "200":
          description: Success
          # schema:
            # $ref: "#/definitions/RoutinesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /users/signup:
    x-swagger-router-controller: users_signup

    post:
      description: POST request for user sign-up
      operationId: usersSignup
      parameters:
        - name: user
          in: body
          description: Create new user account
          required: true
          schema:
            $ref: '#/definitions/Signup'

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UsersResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /users/login:
    x-swagger-router-controller: users_login
    post:
      description: POST request for user to log-in
      operationId: usersLogin
      parameters:
        - name: user
          in: body
          description: Create new user account
          required: true
          schema:
            $ref: '#/definitions/Login'
            
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/UsersResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /users/{users_id}/exercises:
    x-swagger-router-controller: users_id_exercises
    
    get:
      description: GET request for user's custom exercises
      operationId: getUsersIdExercises
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
          description: Input users id
        
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
             $ref: "#/definitions/ExercisesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    post:
      description: POST request for user to create custom exercise
      operationId: postUsersIdExercises
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: exercises
          in: body
          required: true
          schema:
            $ref: '#/definitions/CustomExercise'
            
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ExercisesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
            
  /users/{users_id}/exercises/{id}:
    x-swagger-router-controller: users_id_exercises_id
    
    get:
      description: GET request for specific user's custom routine
      operationId: getUsersIdexercisesId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
          description: Input users id
        - name: id
          in: path
          required: true
          type: integer
          description: Input routines id
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ExercisesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    patch:
      description: PATCH request for user to update custom routine
      operationId: patchUsersIdexercisesId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        - name: routine
          in: body
          required: true
          schema: 
            $ref: '#/definitions/CustomRoutine'
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ExercisesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      description: DELETE request for user's custom routine
      operationId: deleteUsersIdexercisesId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        - name: name
          in: query
          required: true
          type: string
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ExercisesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users/{users_id}/routines:
    x-swagger-router-controller: users_id_routines
    
    get:
      description: GET request for user's custom routine
      operationId: getUsersIdRoutines
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
          description: Input users id
        
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
             $ref: "#/definitions/RoutinesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    post:
      description: POST request for user to create custom routine
      operationId: postUsersIdRoutines
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: routine
          in: body
          required: true
          schema:
            $ref: '#/definitions/CustomRoutine'
            
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RoutinesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
  /users/{users_id}/routines/{id}:
    x-swagger-router-controller: users_id_routines_id
    
    get:
      description: GET request for specific user's custom routine
      operationId: getUsersIdRoutinesId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
          description: Input users id
        - name: id
          in: path
          required: true
          type: integer
          description: Input routines id
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RoutinesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    patch:
      description: PATCH request for user to update custom routine
      operationId: patchUsersIdRoutineId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        - name: routine
          in: body
          required: true
          schema: 
            $ref: '#/definitions/CustomRoutine'
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RoutinesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      description: DELETE request for user's custom routine
      operationId: deleteUsersIdRoutineId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        - name: name
          in: query
          required: true
          type: string
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/RoutinesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users/{users_id}/favorites:
    x-swagger-router-controller: users_id_favorites
    
    get:
      description: GET request for user's favorite routines
      operationId: getUsersIdFavorites
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer

      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/FavoritesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    post:
      description: POST request for user to add routine to favorites
      operationId: postUsersIdFavorites
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
          description: Input user Id
        - name: favorites
          in: body
          required: true
          schema:
            $ref: '#/definitions/Favorites'
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/FavoritesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users/{users_id}/favorites/{id}:
    x-swagger-router-controller: users_id_favorites_id
    
    get:
      description: GET request for user's specifc favorite routine
      operationId: getUsersIdFavoritesId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/FavoritesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      description: DELETE request for user to remove routine from favorites
      operationId: deleteUsersIdFavoritesId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/FavoritesResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users/{users_id}/log:
    x-swagger-router-controller: users_id_log
    
    get:
      description: GET request for array of all log entries
      operationId: getUsersIdLog
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer

      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
             $ref: "#/definitions/LogResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    post:
      description: POST request for user to add routine to log
      operationId: postUsersIdLog
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: log_entry
          in: body
          required: true
          schema:
            $ref: '#/definitions/Log_Entry'
          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LogResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users/{users_id}/log/{id}:
    x-swagger-router-controller: users_id_log_id
    
    get:
      description: GET request for user's specific log entry
      operationId: getUsersIdLogId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LogResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
    patch:
      description: PATCH request for user to update log entry
      operationId: patchUsersIdLogId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        - name: log_entry
          in: body
          required: true
          schema:
            $ref: '#/definitions/Log_Entry'
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LogResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
    delete:
      description: DELETE request for user to remove log entry
      operationId: deleteUsersIdLogId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LogResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users/{users_id}/weight:
    x-swagger-router-controller: users_id_weight
    
    get:
      description: GET request for array of all weight log entries
      operationId: getUsersIdWeight
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: "#/definitions/WeightResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    post:
      description: POST request for user to add weight log entry
      operationId: postUsersIdWeight
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: weight
          in: body
          required: true
          schema:
            $ref: '#/definitions/Weight_Entry'
          
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/WeightResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /users/{users_id}/weight/{id}:
    x-swagger-router-controller: users_id_weight_id
    
    get:
      description: GET request for user's specific weight log entry
      operationId: getUsersIdWeightId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer

      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/WeightResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
    patch:
      description: PATCH request for user to update weight log entry
      operationId: patchUsersIdWeightId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        - name: weight
          in: body
          required: true
          schema:
            $ref: '#/definitions/Weight_Entry'
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/WeightResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  
    delete:
      description: DELETE request for user to remove entry from weight log
      operationId: deleteUsersIdWeightId
      parameters:
        - name: users_id
          in: path
          required: true
          type: integer
        - name: id
          in: path
          required: true
          type: integer
        
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/WeightResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"



  /swagger:
    x-swagger-pipe: swagger_raw
definitions:

  Signup:
    allOf:
      - required: 
        - first_name
        - last_name
        - weight 
        - email
        - password
        properties:
          first_name:
            description: Input first name
            type: string
          last_name: 
            description: Input last name
            type: string
          weight: 
            description: Input user weight (lb)
            type: integer
          email:
            description: Input email
            type: string
          password: 
            description: Input password
            type: string
            format: password
          
  Login:
    allOf:
      - required: 
        - email
        - password
        properties:
          email:
            description: Input email
            type: string
          password:
            description: Input password
            type: string
            
  CustomExercise:
    allOf:
      - required: 
        - name
        - description
        - status
        properties:
          name:
            description: Input name of exercise
            type: string
          description:
            description: Input description of exercise
            type: string
          status:
            description: Input 1-3 ~ 1 = dynamic, 2 = static, 3 = cardio
            type: integer
            minimum: 1
            maximum: 3
            
  CustomRoutine:
    allOf:
      - required: 
        - name
        - description
        - exercises
        properties:
          name:
            description: Input name of routine
            type: string
          description:
            description: Input description of routine
            type: string
          exercises:
            description: Input exercise Ids
            type: array 
            items: 
              type: integer
              
  Log_Entry:
    allOf:
      - required: 
        - routines_id
        - rating
        - notes
        - date
        properties:
          routines_id:
            description: Input routine Id
            type: integer
          rating:
            description: Input rating of completed routine from 1-5
            type: integer
            minimum: 1
            maximum: 5
          notes:
            description: Input notes on completed routine
            type: string
          date:
            description: Input date
            type: string
            format: date

  Weight_Entry:
    allOf:
      - required: 
        - weight
        - date
        properties:
          weight:
            description: Input weight in pounds (lb)
            type: integer
          date:
            description: Input date
            type: string
            format: date
            
  Favorites:
    allOf:
      - required: 
        - routines_id
        properties:
          routines_id:
            description: Input routine Id
            type: integer

  
  UsersResponse:
    required:
      - id
      - first_name
      - last_name
      - weight
      - email
    properties:
        id:
          type: integer
        first_name:
          type: string
        last_name:
          type: string
        weight:
          type: integer
        email:
          type: string
  
  ExercisesResponse:
    type: object
    required:
      - id
      - name
      - description
      - status
    properties:
      id: 
        type: integer
      name:
        type: string
      description:
        type: string
      status:
        type: integer

  RoutinesResponse:
    required:
      - id
      - name
      - description
    properties:
      id: 
        type: integer
      name:
        type: string
      description:
        type: string

  FavoritesResponse:
    required:
      - id
      - users_id
      - routines_id
    properties:
      id:
        type: integer
      users_id: 
        type: integer
      routines_id:
        type: integer
  
  LogResponse: 
    required:
      - id
      - users_id
      - routines_id
      - rating
      - notes
      - date
    properties:
      id:
        type: integer
      users_id: 
        type: integer
      routines_id:
        type: integer
      rating:
        type: integer
      notes:
        type: string
      date:
        type: string

  WeightResponse:
    required:
      - id
      - users_id
      - weight
      - date
    properties:
      id:
        type: integer
      users_id: 
        type: integer
      weight:
        type: integer
      date:
        type: string
  
  ErrorResponse:
    properties:
      ErrorMessage:
        type: string